Cherry-picked 2d02d6ef
* Use IValueSetFacade instead of the concrete type
* Deduped code for creating ValueSetFacades
* Simplified ValueSetUnion to use the IValueSetDictionary of the facade.

The only remotely interesting point here is the third bullet. For some reason that I cannot grasp. the ValueSet union had a constructor which took an IValueSetDictionary which it will use to look up the codes in the IValueSetFacades to join. That is confusing, since these facades themselves already have a reference to their IValueSetDictionary. It seems a source of errors if these become incompatible, and also it seems to make sense that the ValueSetFacades "know best", so I changed the code so the valueset facades will provide the codes themselves to the union, instead of them being taken from another (possibly inconsistent) reference to IValueSetDictionary.